refactor toast components
