+++ src/betterdiscord/ui/toasts/ToastContainer.tsx
@@ -1,11 +1,22 @@
 import React, {Fragment} from "@modules/react";
 import DiscordModules from "@modules/discordmodules";
+import ToastIcon from "@ui/toasts/ToastIcon";
 import ToastStore from "@stores/toasts";
 import {useInternalStore} from "@ui/hooks";
 
-import Toast, {type ToastType} from "@ui/toasts/Toast";
+import clsx from "clsx";
+import type {AnimatedProps} from "@react-spring/web";
 
 const ReactSpring = DiscordModules.ReactSpring;
+
+export type ToastType = "default" | "info" | "success" | "warning" | "error";
+
+interface ToastItemProps {
+    content: string;
+    type: ToastType;
+    icon: boolean;
+    style: AnimatedProps<React.CSSProperties>;
+}
 
 export interface ToastProps {
     key: number;
@@ -13,6 +24,13 @@ export interface ToastProps {
     type: ToastType;
     icon: boolean;
     timeout: number;
+}
+
+export function Toast({content, type, icon, style}: ToastItemProps) {
+    return <ReactSpring.animated.div className={clsx("bd-toast", `toast-${type}`)} style={style}>
+        {icon && <ToastIcon type={type} />}
+        <span>{content}</span>
+    </ReactSpring.animated.div>;
 }
 
 export default function ToastContainer() {
@@ -27,14 +45,14 @@ export default function ToastContainer() {
     });
 
     return <Fragment>
-            {transition((style, item) => (
-                <Toast
-                    key={item.key}
-                    content={item.content}
-                    type={item.type}
-                    icon={item.icon}
-                    style={style}
-                />
-            ))}
-        </Fragment>;
+        {transition((style, item) => (
+            <Toast
+                key={item.key}
+                content={item.content}
+                type={item.type}
+                icon={item.icon}
+                style={style}
+            />
+        ))}
+    </Fragment>;
 }
